fix: move data class in class
