+++ piikii-application/src/main/kotlin/com/piikii/application/port/input/roomcategory/dto/request/RoomCategoriesCreateRequest.kt
@@ -1,5 +1,6 @@
 package com.piikii.application.port.input.roomcategory.dto.request
 
+import com.piikii.application.domain.roomcategory.PlaceCategory
 import com.piikii.application.domain.roomcategory.RoomCategory
 import io.swagger.v3.oas.annotations.media.Schema
 import java.util.UUID
@@ -10,5 +11,23 @@ data class RoomCategoriesCreateRequest(
 ) {
     fun toDomains(roomId: UUID): List<RoomCategory> {
         return roomCategories.map { it.toDomain(roomId) }
+    }
+}
+
+data class RoomCategoryCreateRequest(
+    @Schema(description = "방 카테고리의 이름")
+    val name: String,
+    @Schema(description = "방 카테고리의 카테고리")
+    val category: PlaceCategory,
+    @Schema(description = "방 카테고리의 순서")
+    val sequence: Int,
+) {
+    fun toDomain(roomId: UUID): RoomCategory {
+        return RoomCategory(
+            roomId = roomId,
+            name = this.name,
+            category = this.category,
+            sequence = this.sequence,
+        )
     }
 }
